CLN: Moved MultiImage subclasses to pyfibre.model.objects module
